Add square collision detection
Added detect_collision which checks the square area around the marble and compares those positions to map. Right now it only output to the cmd window if it detects a collision. It also alerts if the bounds of the marble fall outside the 600x600 plane. Still need to change from a square area to check to the actual circle

Changed update_tick so it enforces integers only for position.

Added a feature that kills the script if the marble falls outside of the 600x600 plane

Changed radius for draw function to 5

Changed accelerations in main for testing