Add Basic Physics
The ball now (usually) doesn't pass through solids. It moves down slopes. X acceleration just gets inverted to make this happen. Y acceleration is gravity/mass when its not in contact with a surface below the sphere. These "physics" are terrible but purely for testing and demonstration for now.

Increased time step to 100 per second so it is less likely for the sphere to "miss" a solid object because it was moving too fast.

Moved the pause function to main and only runs every 5th tick so the program can actually finish in a decent time

Changed the division factor in update_tick